1. Start discovery process         2. Change Outputdevice struct         3. Start cast process and listener, modified provider map to providerManager map         4. CastController, play_info_holder and play_info         5. Others         6. New interface_sdk_js         7. Ready to merge, add callbacks         8. Fix cast controller getter
        modified:   frameworks/common/src/avcast_control_command.cpp
        modified:   frameworks/js/napi/session/include/napi_avcast_controller.h
        modified:   frameworks/js/napi/session/include/napi_avcast_controller_callback.h
        modified:   frameworks/js/napi/session/src/napi_avcast_controller.cpp
        modified:   frameworks/js/napi/session/src/napi_avcast_controller_callback.cpp
        modified:   frameworks/js/napi/session/src/napi_cast_control_command.cpp
        modified:   frameworks/native/session/include/avcast_controller_callback_client.h
        modified:   frameworks/native/session/src/avcast_controller_callback_client.cpp
        modified:   interfaces/inner_api/native/session/include/avcast_control_command.h
        modified:   interfaces/inner_api/native/session/include/avcast_controller.h
        modified:   interfaces/inner_api/native/session/include/avsession_info.h
        modified:   interfaces/inner_api/native/session/include/i_avcast_controller_proxy.h
        modified:   services/session/ipc/base/iavcast_controller.h
        modified:   services/session/ipc/proxy/avcast_controller_callback_proxy.cpp
        modified:   services/session/ipc/proxy/avcast_controller_callback_proxy.h
        modified:   services/session/ipc/proxy/avcast_controller_proxy.cpp
        modified:   services/session/ipc/proxy/avcast_controller_proxy.h
        modified:   services/session/ipc/stub/avcast_controller_callback_stub.cpp
        modified:   services/session/ipc/stub/avcast_controller_stub.cpp
        modified:   services/session/ipc/stub/avcast_controller_stub.h
        modified:   services/session/server/avcast_controller_item.cpp
        modified:   services/session/server/avcast_controller_item.h
        modified:   utils/include/avsession_trace.h

+++ frameworks/common/src/avcast_control_command.cpp
@@ -42,7 +42,7 @@ AVCastControlCommand *AVCastControlCommand::Unmarshalling(Parcel& data)
                 result->SetVolume(data.ReadInt32());
                 break;
             case CAST_CONTROL_CMD_SET_SPEED:
-                result->SetSpeed(data.ReadDouble());
+                result->SetSpeed(data.ReadInt32());
                 break;
             default:
                 break;
@@ -66,8 +66,8 @@ bool AVCastControlCommand::Marshalling(Parcel& parcel) const
                 && parcel.WriteInt32(std::get<int32_t>(param_)), false, "write volume failed");
             break;
         case CAST_CONTROL_CMD_SET_SPEED:
-            CHECK_AND_RETURN_RET_LOG(std::holds_alternative<double>(param_)
-                && parcel.WriteDouble(std::get<double>(param_)), false, "write speed failed");
+            CHECK_AND_RETURN_RET_LOG(std::holds_alternative<int32_t>(param_)
+                && parcel.WriteInt32(std::get<int32_t>(param_)), false, "write speed failed");
             break;
         default:
             break;
@@ -94,7 +94,7 @@ int32_t AVCastControlCommand::GetCommand() const
     return cmd_;
 }
 
-int32_t AVCastControlCommand::SetSpeed(double speed)
+int32_t AVCastControlCommand::SetSpeed(int32_t speed)
 {
     if (speed <= 0) {
         return ERR_INVALID_PARAM;
@@ -103,9 +103,9 @@ int32_t AVCastControlCommand::SetSpeed(double speed)
     return AVSESSION_SUCCESS;
 }
 
-int32_t AVCastControlCommand::GetSpeed(double& speed) const
+int32_t AVCastControlCommand::GetSpeed(int32_t& speed) const
 {
-    if (!std::holds_alternative<double>(param_)) {
+    if (!std::holds_alternative<int32_t>(param_)) {
         return AVSESSION_ERROR;
     }
     speed = std::get<int32_t>(param_);
