+++ services/session/ipc/proxy/avcast_controller_callback_proxy.cpp
@@ -79,11 +79,11 @@ void AVCastControllerCallbackProxy::OnLoopModeChange(const int32_t loopMode)
         "send request failed");
 }
 
-void AVCastControllerCallbackProxy::OnPlaySpeedChange(const double playSpeed)
+void AVCastControllerCallbackProxy::OnPlaySpeedChange(const int32_t playSpeed)
 {
     MessageParcel parcel;
     CHECK_AND_RETURN_LOG(parcel.WriteInterfaceToken(GetDescriptor()), "write interface token failed");
-    CHECK_AND_RETURN_LOG(parcel.WriteDouble(playSpeed), "write speed failed");
+    CHECK_AND_RETURN_LOG(parcel.WriteInt32(playSpeed), "write speed failed");
 
     MessageParcel reply;
     MessageOption option = { MessageOption::TF_ASYNC };
